Consolidate strain-trajectory trellises into one notebook
Merge the separate full-length and zoom notebooks into a single
visualization notebook (no CLI args) that renders four trellis sets from
one make_trellis helper: the full run (all 5000 updates, all power-of-ten
rates incl. 1e-9) plus establishment-phase zooms to the first 2000, 400,
and 200 updates (which drop the slowest 1e-9 rate). Remove the
now-redundant -zoom notebook.
